[refactor] 컴포넌트 분리, 타입 추가, 주석 제거
