+++ app/_api/constant.ts
@@ -9,7 +9,7 @@ export const MODE_DETAIL_PAGE = "detailPage";
 export const MODE_INDIVIDUAL_ACTION = "individualAction";
 export const MODE_INDIVIDUAL_ACTION_ADD = "individualAdd";
 export const MODE_COMMENT = "comment";
-export const MODE_ADD_COMMENT = "addComment"
+export const MODE_ADD_COMMENT = "addComment";
 export const MODE_ADD_ACTION = "addAction";
 export const MODE_ADD_POST = "addPost";
 export const MODE_IN = "in";
@@ -23,3 +23,4 @@ export const MODE_MOBILE = "mobile";
 export const ACTIVE_TABS = "단체와 함께해요";
 export const MODE_TODAY = "todayAlertRoomIds";
 export const MODE_PREVIOUS = "previousAlertRoomIds";
+export const ACTIVE_TAB_MY_ACTION = "나의 Green-Action";
