+++ app/_components/main/LaptopMainSlidder.tsx
@@ -11,15 +11,12 @@ import CommunityListPost from "../community/CommunityListPost";
 import MyActionCard from "../mypage/MyActionCard";
 import SoomLoaing from "/app/_assets/image/loading/SOOM_gif.gif";
 
-// export const revalidate = 0;
-
 const LaptopMainSlidder = ({ mode }: { mode: string }) => {
   // FIXME 메인페이지 mode action 인 경우 모집중인 것만 뜨게할지?
 
   var settings = {
     autoplay: true,
     autoplaySpeed: 15000,
-    // fade: true,
     dots: true,
     infinite: true,
     arrows: false,
@@ -78,7 +75,7 @@ const LaptopMainSlidder = ({ mode }: { mode: string }) => {
               action, // 북마크 수 최다 상위 8개 action
             ) => (
               <div key={action.id} className="flex items-center h-[480px]">
-                <MyActionCard action={action} mode={MODE_MAIN} />
+                <MyActionCard action={action as any} mode={MODE_MAIN} />
               </div>
             ),
           )}
