+++ app/_types/mypage/mypage.ts
@@ -3,13 +3,13 @@ export interface MyAction {
   content: string | null;
   created_at: string;
   end_date: string | null;
-  is_recruiting?: boolean;
+  is_recruiting: boolean;
   location: string | null;
   recruit_number: number | null;
   start_date: string | null;
-  title: string | null;
+  title: string;
   user_uid: string | null;
-  actionImgUrls: string[];
+  actionImgUrls: { img_url: string }[];
   bookmarkedAction: {
     id: string;
     content: string | null;
@@ -19,12 +19,38 @@ export interface MyAction {
     location: string | null;
     recruit_number: number | null;
     start_date: string | null;
-    title: string | null;
+    title: string;
     user_uid: string | null;
-    actionImgUrls: string[];
+    actionImgUrls: { img_url: string }[];
   };
 }
 
+export interface BookmarkedAction {
+  id: string;
+  content: string | null;
+  created_at: string;
+  end_date: string | null;
+  is_recruiting: boolean;
+  location: string | null;
+  recruit_number: number | null;
+  start_date: string | null;
+  title: string;
+  user_uid: string | null;
+  actionImgUrls: { img_url: string }[];
+}
+
+export interface actionImgUrls {
+  actionImgUrls: { img_url: string }[];
+}
+
+export interface userInfoProps {
+  display_name: string | null;
+  email: string | null;
+  id: string;
+  introduction: string | null;
+  point: number | null;
+  profile_img: string | null;
+}
 export interface ProfileFileUpload {
   file: File | undefined;
   user_uid: string;
