+++ app/mypage/page.tsx
@@ -1,54 +1,24 @@
 "use client";
 
+import React from "react";
 import { useSession } from "next-auth/react";
 import Image from "next/image";
-import { useRouter } from "next/navigation";
-import React, { useEffect, useState } from "react";
-
-import {
-  useFetchMyGreenActions,
-  useFetchUserInfo,
-  usefetchBookmarkedActions,
-  usefetchMyCommunityPosts,
-} from "../_hooks/useQueries/mypage";
-import { User } from "../_types";
-
-import CommunityListPost from "../_components/community/CommunityListPost";
-import MyActionCard from "../_components/mypage/MyActionCard";
+import { useFetchUserInfo } from "../_hooks/useQueries/mypage";
+import { useResponsive } from "../_hooks/responsive";
 import MyProfile from "../_components/mypage/MyProfile";
-import RecruitSelectTab from "../_components/mypage/RecruitSelectTab";
-
 import TopButton from "../_components/TopButton";
 import MyActionCardMobile from "../_components/mypage/MyActionCardMobile";
-import { useResponsive } from "../_hooks/responsive";
 import SoomLoading from "/app/_assets/image/loading/SOOM_gif.gif";
 
-const MyPage = () => {
-  // TODO props 타입등 재설정
-  const router = useRouter();
-  const session = useSession();
-  const { isDesktop, isLaptop, isMobile } = useResponsive();
+import type { User } from "../_types";
+import MyPageList from "../_components/mypage/MyPageList";
+import { userInfoProps } from "../_types/mypage/mypage";
 
-  const isLoggedIn = !!session.data;
+const MyPage = () => {
+  const session = useSession();
   const user_uid = session.data?.user.user_uid as string;
 
-  const {
-    data: myActions,
-    isLoading: isActionsLoading,
-    isError: isActionsError,
-  } = useFetchMyGreenActions(user_uid);
-
-  const {
-    data: myPosts,
-    isLoading: isPostsLoading,
-    isError: isPostsError,
-  } = usefetchMyCommunityPosts(user_uid);
-
-  const {
-    data: myBookmarks,
-    isLoading: isBookmarksLoading,
-    isError: isBookmarksError,
-  } = usefetchBookmarkedActions(user_uid);
+  const { isDesktop, isLaptop, isMobile } = useResponsive();
 
   // 유저 정보 조회
   const {
@@ -57,92 +27,7 @@ const MyPage = () => {
     isError: isUserInfoError,
   } = useFetchUserInfo(user_uid);
 
-  const { display_name, profile_img } = (userInfo as User["userInfo"]) || ""; // as User["userInfo"] 외에도 || '' 처리해줘야 에러안뜸
-
-  // my action - created_at (작성일) 기준으로 정렬하기
-  const sortedMyActions = myActions?.slice().sort((a, b) => {
-    return new Date(a.created_at).getTime() - new Date(b.created_at).getTime();
-  });
-
-  // 찜한 action - created_at (작성일) 기준으로 정렬하기
-  const sortedMyBookmarks = myBookmarks?.slice().sort((a, b) => {
-    return (
-      new Date(a?.bookmarkedAction?.created_at || "").getTime() -
-      new Date(b?.bookmarkedAction?.created_at || "").getTime()
-    );
-  });
-
-  const [activeTab, setActiveTab] = useState("나의 Green-Action");
-  const [myRecruitClicked, setMyRecruitClicked] = useState("전체");
-  const [bookmarkedRecruitClicked, setBookmarkedRecruitClicked] =
-    useState("전체");
-  const [filteredActions, setFilteredActions] = useState(sortedMyActions);
-  const [filteredBookmarkedActions, setFilteredBookmarkedActions] =
-    useState(sortedMyBookmarks);
-
-  useEffect(() => {
-    filterByRecruiting();
-  }, [
-    myActions,
-    myBookmarks,
-    activeTab,
-    myRecruitClicked,
-    bookmarkedRecruitClicked,
-  ]);
-
-  // My Action, 작성 커뮤니티 글, 찜한 Action 탭 선택시
-  const handleActiveTabClick = (
-    e: React.MouseEvent<HTMLDivElement, MouseEvent>,
-    // e: React.MouseEvent<HTMLButtonElement, MouseEvent>,
-  ) => {
-    const target = e.target as HTMLButtonElement;
-    const textContent = target.textContent;
-    if (textContent) {
-      setActiveTab(textContent);
-    }
-  };
-
-  const filterByRecruiting = () => {
-    if (activeTab === "나의 Green-Action") {
-      if (myRecruitClicked === "전체") {
-        setFilteredActions(sortedMyActions);
-      }
-      if (myRecruitClicked === "모집 중") {
-        setFilteredActions(
-          sortedMyActions?.filter((action) => action.is_recruiting),
-        );
-      } else if (myRecruitClicked === "모집 마감") {
-        setFilteredActions(
-          sortedMyActions?.filter((action) => !action.is_recruiting),
-        );
-      }
-    }
-    if (activeTab === "즐겨찾는 Green-Action") {
-      if (bookmarkedRecruitClicked === "전체") {
-        setFilteredBookmarkedActions(sortedMyBookmarks);
-      }
-      if (bookmarkedRecruitClicked === "모집 중") {
-        setFilteredBookmarkedActions(
-          sortedMyBookmarks?.filter(
-            (action) => action.bookmarkedAction?.is_recruiting,
-          ),
-        );
-      } else if (bookmarkedRecruitClicked === "모집 마감") {
-        setFilteredBookmarkedActions(
-          sortedMyBookmarks?.filter(
-            (action) => !action.bookmarkedAction?.is_recruiting,
-          ),
-        );
-      }
-    }
-  };
-
-  if (
-    isActionsLoading ||
-    isPostsLoading ||
-    isBookmarksLoading ||
-    isUserInfoLoading
-  ) {
+  if (isUserInfoLoading) {
     return (
       <div className="flex justify-center items-center w-screen h-[500px]">
         <Image
@@ -155,7 +40,7 @@ const MyPage = () => {
     );
   }
 
-  if (isActionsError || isPostsError || isBookmarksError || isUserInfoError) {
+  if (isUserInfoError) {
     return (
       <div className="flex justify-center items-center w-screen h-[500px]">
         ❌ ERROR : 이 페이지를 표시하는 도중 문제가 발생했습니다. 다른 페이지로
@@ -177,97 +62,7 @@ const MyPage = () => {
           <div className="flex flex-col desktop:pl-[82px] laptop:pl-[30px] desktop:pt-1 laptop:pt-[30px] w-full">
             {isMobile && <MyProfile userInfo={userInfo as User["userInfo"]} />}
             {(isDesktop || isLaptop) && (
-              <>
-                <div className="flex justify-between laptop:mb-[30px] vh-auto">
-                  <div className="flex desktop:gap-[45px] laptop:gap-[30px] desktop:ml-5 desktop:text-[12pt]">
-                    <div
-                      onClick={handleActiveTabClick}
-                      className={`cursor-pointer h-[30px] desktop:text-[12pt] laptop:text-[11pt] 
-                    ${
-                      activeTab === "나의 Green-Action" &&
-                      "border-b-2 border-[#979797] transition duration-400 ease-in-out"
-                    }`}
-                    >
-                      나의 Green-Action
-                    </div>
-                    <div
-                      onClick={handleActiveTabClick}
-                      className={`cursor-pointer h-[30px] desktop:text-[12pt] laptop:text-[11pt]
-                    ${
-                      activeTab === "즐겨찾는 Green-Action" &&
-                      "border-b-2 border-[#979797] transition duration-400 ease-in-out"
-                    }`}
-                    >
-                      즐겨찾는 Green-Action
-                    </div>
-                    <div
-                      onClick={handleActiveTabClick}
-                      className={`cursor-pointer h-[30px] desktop:text-[12pt] laptop:text-[11pt]
-                   ${
-                     activeTab === "나의 Community" &&
-                     "border-b-2 border-[#979797] transition duration-400 ease-in-out"
-                   }`}
-                    >
-                      나의 Community
-                    </div>
-                  </div>
-                  <div className="mr-[80px]">
-                    {activeTab === "나의 Green-Action" && (
-                      <RecruitSelectTab
-                        selected={myRecruitClicked}
-                        setSelected={setMyRecruitClicked}
-                      />
-                    )}
-                    {activeTab === "즐겨찾는 Green-Action" && (
-                      <RecruitSelectTab
-                        selected={bookmarkedRecruitClicked}
-                        setSelected={setBookmarkedRecruitClicked}
-                      />
-                    )}
-                  </div>
-                </div>
-                <div className="flex flex-wrap gap-[20px]">
-                  {/* LINK My Green Action */}
-                  {activeTab === "나의 Green-Action" &&
-                    filteredActions?.map((action) => {
-                      return (
-                        <MyActionCard
-                          key={action.id}
-                          action={action}
-                          mode="myPosts"
-                        />
-                      );
-                    })}
-                </div>
-                {/* LINK 찜한 Green Action */}
-                <div className="flex flex-wrap gap-[20px]">
-                  {activeTab === "즐겨찾는 Green-Action" &&
-                    filteredBookmarkedActions?.map((bookmark) => {
-                      return (
-                        <MyActionCard
-                          key={bookmark?.bookmarkedAction?.id || ""}
-                          action={bookmark}
-                          mode="myBookmarks"
-                        />
-                      );
-                    })}
-                </div>
-                {/* LINK 내가 쓴 커뮤니티 글 */}
-                <div className="flex flex-wrap gap-[20px]">
-                  {activeTab === "나의 Community" &&
-                    myPosts?.map((post) => {
-                      return (
-                        <CommunityListPost
-                          key={post.id}
-                          mode="myPosts"
-                          communityPost={post}
-                          my_display_name={display_name}
-                          my_profile_img={profile_img || null}
-                        />
-                      );
-                    })}
-                </div>
-              </>
+              <MyPageList userInfo={userInfo as userInfoProps} />
             )}
             {isMobile && <MyActionCardMobile />}
           </div>
