+++ app/not-found.tsx
@@ -5,7 +5,6 @@ const NotFoundPage = () => {
       <p>:경고: 잘못된 접근입니다. 이 페이지는 존재하지 않습니다. :x:</p>
       <p>삭제되었거나 잘못된 주소입니다.</p>
     </div>
-    // 나중에 바꾸는 걸로
   );
 };
 export default NotFoundPage;
